requirements.txt
python-telegram-bot==21.3
openai==1.35.3
python-dotenv
fastapi
uvicorn[standard]
jinja2
itsdangerous
httpx
